Moved acting agent to abs
